+++ CHANGELOG.md
@@ -878,6 +878,7 @@
 ## [unreleased]
 
 - BEIS users create reports manually
+- Create historical events when creating a refund
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-81...HEAD
 [release-81]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-80...release-81